code/requirements.txt
sentence-transformers==3.4.1
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
